Moved all from utils to utils.misc
