+++ bukkit-utils/src/main/java/net/anweisen/utilities/bukkit/utils/misc/MinecraftVersion.java
@@ -1,4 +1,4 @@
-package net.anweisen.utilities.bukkit.utils;
+package net.anweisen.utilities.bukkit.utils.misc;
 
 import net.anweisen.utilities.commons.version.Version;
